requirements.txt
psutil
pyyaml
rich
